:recycle: change Integer id to Long id

+++ src/main/java/com/twd/technicaltest/application/usecase/DeleteHotelUseCase.java
@@ -1,5 +1,5 @@
 package com.twd.technicaltest.application.usecase;
 
 public interface DeleteHotelUseCase {
-    void  delete(Integer id);
+    void  delete(Long id);
 }
